Add VIF Function and improve admin name extraction
The VIF function selects variables that have a VIF value that is below
a threshold specified by the user. The variables are removed
iterartively, starting from the one with the highest VIF, recalculating
the VIF values, then removing the new Highest. The process stops when
the highest VIF is lower than the threshold.

The process of extracting Admin names is still slow, however, the user
can choose to skip it and use a dataset that contains admin values.