Version 1.0.0
Fully implemented the model selection process. The process results is
saved so that the step can be skipped - it takes 3 hours.

I made a couple of changes before the model selection step:
(i) I reintroduced fertilizer use in place of fertilizer amount because about
half the values of fertilizer amount were zero
(ii) I changed maize area to binary: less than 1Ha or greater than or equal
to 1Ha - about half the values are less than 1Ha
(iii) After categorizing distance to market I noticed that the values were only significant in two categories: 30min to 35 min and above 35minutes. So I changed to binary: less than 30 minutes or greater than or equal to 30 minutes
(iv) The response variable(maize produce in kg/Ha) is left-skewed, but I just
standardized it by subtracting the mean and dividing it by the standard
deviation.
(v)I categorized distance to extension services, but it is not
selected in the optimal model in the attachment below
ssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssssss